agent/requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.3
langchain-tavily>=0.2.0
tavily-python>=0.3.3
requests>=2.31.0
python-dotenv>=1.0.1
